(loadSimTree) Added duplicated variable check. Enabled removal of rows from table.

+++ tests/gui/views/simulation_tree.py
@@ -1,7 +1,7 @@
 import sys
 from PyQt5.QtWidgets import QDialog, QApplication
 from PyQt5.QtGui import QStandardItemModel
-from PyQt5.QtCore import Qt
+from PyQt5.QtCore import Qt, QEvent
 from gui.views.loadSimulationTree import *
 from gui.models.load_simulation import construct_tree_items, read_simulation_tree
 
@@ -14,8 +14,8 @@ class LoadSimulationTreeDialog(QDialog):
 
 		self.ui.pushButtonOK.clicked.connect(self.close)  # closes window when button is pressed
 
-		self.ui.treeViewInput.doubleClicked.connect(self.treedoubleclick)  # assign double click event
-		self.ui.treeViewOutput.doubleClicked.connect(self.treedoubleclick)
+		self.ui.treeViewInput.doubleClicked.connect(self.treeDoubleClick)  # assign double click event
+		self.ui.treeViewOutput.doubleClicked.connect(self.treeDoubleClick)
 
 		self.ui.tableWidgetInput.setColumnWidth(0, 550)  # first column resizing
 		self.ui.tableWidgetOutput.setColumnWidth(0, 550)
@@ -48,7 +48,16 @@ class LoadSimulationTreeDialog(QDialog):
 		model_tree_input.appendRow(blocks_input)
 		model_tree_output.appendRow(blocks_output)
 
-	def treedoubleclick(self):
+	def keyPressEvent(self, event):
+		if event.type() == QEvent.KeyPress and event.key() == Qt.Key_Delete:  # if delete key was pressed
+			if self.ui.tableWidgetInput.hasFocus():  # and is inside the table widget
+				table_model = self.ui.tableWidgetInput.selectionModel()
+				indices = table_model.selectedRows()
+
+				for index in indices:
+					self.ui.tableWidgetInput.removeRow(index.row())  # delete selected rows
+
+	def treeDoubleClick(self):
 		"""
 		Event handling associated with a double click of a single LEAF node of the tree
 		"""
@@ -81,15 +90,41 @@ class LoadSimulationTreeDialog(QDialog):
 			# create full path name
 			branch_str = '\\' + '\\'.join(list(reversed(branch_list)))
 
+			# verify if the entry is already in the table
 			row_position = table_view.rowCount()  # count the current number of rows
-			table_view.insertRow(row_position)  # insert an empty row
-			table_item_path = QtWidgets.QTableWidgetItem(branch_str)
-			table_view.setItem(row_position, 0, table_item_path)
-			table_view.setItem(row_position, 1, QtWidgets.QTableWidgetItem('Alias_' + str(row_position)))
-			flags = Qt.ItemFlags()
-			flags != Qt.ItemIsEditable
 
-			table_item_path.setFlags(flags)  # disable edit of the first col
+			if row_position != 0:  # table is not empty
+				row_list = []
+				for i in range(row_position):
+					row_list.append(table_view.model().index(i, 0).data())  # get all rows in table
+
+				if branch_str not in row_list:  # there isn't the value in table. Insert it
+					table_view.insertRow(row_position)
+
+					table_item_path = QtWidgets.QTableWidgetItem(branch_str)
+
+					table_item_path.setFlags(Qt.ItemIsSelectable | Qt.ItemIsEnabled)  # disable edit of the first col
+					table_view.setItem(row_position, 0, table_item_path)
+					table_view.setItem(row_position, 1, QtWidgets.QTableWidgetItem('Alias_' + str(row_position)))
+
+				else:
+					# warn the user
+					msg_box = QtWidgets.QMessageBox()
+					msg_box.setIcon(QtWidgets.QMessageBox.Warning)
+					msg_box.setText("The selected variable is already the table!")
+					msg_box.setWindowTitle("Duplicated variable")
+					msg_box.setStandardButtons(QtWidgets.QMessageBox.Ok)
+
+					msg_box.exec()
+			else:
+				table_view.insertRow(row_position)  # insert an empty row
+
+				table_item_path = QtWidgets.QTableWidgetItem(branch_str)
+
+				table_item_path.setFlags(Qt.ItemIsSelectable | Qt.ItemIsEnabled)  # disable edit of the first col
+				table_view.setItem(row_position, 0, table_item_path)
+				table_view.setItem(row_position, 1, QtWidgets.QTableWidgetItem('Alias_' + str(row_position)))
+
 
 
 if __name__ == '__main__':
